modify problems order
