+++ src/class21/Code02_CoinsWayEveryPaperDifferent.java
@@ -1,44 +1,32 @@
 package class21;
 
-public class Code05_MinCoinsNoLimit {
+public class Code02_CoinsWayEveryPaperDifferent {
 
-	public static int minCoins(int[] arr, int aim) {
+	public static int coinWays(int[] arr, int aim) {
 		return process(arr, 0, aim);
 	}
 
 	public static int process(int[] arr, int index, int rest) {
 		if (rest < 0) {
-			return Integer.MAX_VALUE;
+			return 0;
 		}
 		if (index == arr.length) {
-			return rest == 0 ? 0 : Integer.MAX_VALUE;
+			return rest == 0 ? 1 : 0;
 		} else {
-			int ans = Integer.MAX_VALUE;
-			for (int zhang = 0; zhang * arr[index] <= rest; zhang++) {
-				int next = process(arr, index + 1, rest - zhang * arr[index]);
-				if (next != Integer.MAX_VALUE) {
-					ans = Math.min(ans, next + zhang);
-				}
-			}
-			return ans;
+			return process(arr, index + 1, rest) + process(arr, index + 1, rest - arr[index]);
 		}
 	}
 
 	public static int dp(int[] arr, int aim) {
 		if (aim == 0) {
-			return 0;
+			return 1;
 		}
 		int N = arr.length;
 		int[][] dp = new int[N + 1][aim + 1];
-		dp[N][0] = 0;
-		for (int j = 1; j <= aim; j++) {
-			dp[N][j] = Integer.MAX_VALUE;
-		}
+		dp[N][0] = 1;
 		for (int index = N - 1; index >= 0; index--) {
 			for (int rest = 0; rest <= aim; rest++) {
-				dp[index][rest] = dp[index + 1][rest];
-				if (rest - arr[index] >= 0 && dp[index][rest - arr[index]] != Integer.MAX_VALUE)
-					dp[index][rest] = Math.min(dp[index][rest], dp[index][rest - arr[index]] + 1);
+				dp[index][rest] = dp[index + 1][rest] + (rest - arr[index] >= 0 ? dp[index + 1][rest - arr[index]] : 0);
 			}
 		}
 		return dp[0][aim];
@@ -48,12 +36,8 @@ public class Code05_MinCoinsNoLimit {
 	public static int[] randomArray(int maxLen, int maxValue) {
 		int N = (int) (Math.random() * maxLen);
 		int[] arr = new int[N];
-		boolean[] has = new boolean[maxValue + 1];
 		for (int i = 0; i < N; i++) {
-			do {
-				arr[i] = (int) (Math.random() * maxValue) + 1;
-			} while (has[arr[i]]);
-			has[arr[i]] = true;
+			arr[i] = (int) (Math.random() * maxValue) + 1;
 		}
 		return arr;
 	}
@@ -70,13 +54,12 @@ public class Code05_MinCoinsNoLimit {
 	public static void main(String[] args) {
 		int maxLen = 20;
 		int maxValue = 30;
-		int testTime = 300000;
-		System.out.println("功能测试开始");
+		int testTime = 1000000;
+		System.out.println("测试开始");
 		for (int i = 0; i < testTime; i++) {
-			int N = (int) (Math.random() * maxLen);
-			int[] arr = randomArray(N, maxValue);
+			int[] arr = randomArray(maxLen, maxValue);
 			int aim = (int) (Math.random() * maxValue);
-			int ans1 = minCoins(arr, aim);
+			int ans1 = coinWays(arr, aim);
 			int ans2 = dp(arr, aim);
 			if (ans1 != ans2) {
 				System.out.println("Oops!");
@@ -87,7 +70,7 @@ public class Code05_MinCoinsNoLimit {
 				break;
 			}
 		}
-		System.out.println("功能测试结束");
+		System.out.println("测试结束");
 	}
 
 }
