+++ src/class22/Code04_MinCoinsOnePaper.java
@@ -1,4 +1,4 @@
-package class21;
+package class22;
 
 import java.util.HashMap;
 import java.util.Map.Entry;
