+++ src/class23/Code01_SplitSumClosed.java
@@ -1,8 +1,8 @@
-package class22;
+package class23;
 
 import java.util.TreeSet;
 
-public class Code03_SplitSumClosed {
+public class Code01_SplitSumClosed {
 
 	public static int right(int[] arr) {
 		if (arr == null || arr.length < 2) {
